feat :더보기 버튼 클릭 시 상태 유지를 custom hook에서 실행
