+++ src/app/(detail)/components/GBItemDetail/GBItemDetail.tsx
@@ -9,6 +9,8 @@ import GBItemCarousel from '@/app/(detail)/components/GBItemCarousel';
 import GBItemInformation from '@/app/(detail)/components/GBItemInformation';
 import { getProductsDetailQueryObject } from '@/app/(queries)/productsQueries';
 
+import { useUpdateProductOptions } from '../../hooks/useUpdateProductOptions';
+
 import styles from './GBItemDetail.module.scss';
 
 const cx = classNames.bind(styles);
@@ -17,6 +19,8 @@ const GBItemDetail = () => {
   const pathname = usePathname();
 
   const { data } = useSuspenseQuery(getProductsDetailQueryObject(pathname.replace('/', '')));
+
+  useUpdateProductOptions(data.data.productStatus, data.data.productType);
 
   if (!data) {
     return null;
